fix size and color bugs
