+++ brushes.py
@@ -6,7 +6,7 @@ from PyQt5.QtWidgets import *
 import sys
 
 class Brush:
-    def __init__(self, size = 2, color=Qt.black):
+    def __init__(self, size=2, color=Qt.black):
         self.size = size
         self.color = color
 
@@ -16,25 +16,27 @@ class Brush:
     def set_size(self, size):
         self.size = size
 
+
 class BrushSizeInput(QtWidgets.QWidget):
-    def __init__(self, brush, parent=None):
-        super().__init__(parent)
+    def __init__(self, brush, canvas):
+        super().__init__()
+        self.canvas = canvas
         self.brush = brush
         self.brush_size_input = QtWidgets.QLineEdit(self)
         self.brush_size_input.setPlaceholderText("Brush size")
         self.brush_size_input.setGeometry(10, 10, 80, 20)
         self.brush_size_input.returnPressed.connect(self.set_brush_radius_from_input)
 
+        self.color_button = QtWidgets.QPushButton("Choose color", self)
+        self.color_button.clicked.connect(self.choose_color)
+        self.canvas.brush = self.brush
+
     @pyqtSlot(int)
     def update_brush_size(self, new_size):
         self.brush_size_input.setText(str(new_size))
         self.brush.set_size(new_size)
-
-    def choose_color(self):
-        color = QtWidgets.QColorDialog.getColor()
-        if color.isValid():
-            self.brush.color = color
-
+        self.canvas.brush.size = new_size
+        
     def set_brush_radius_from_input(self):
         try:
             self.brush.size = int(self.brush_size_input.text())
@@ -42,6 +44,11 @@ class BrushSizeInput(QtWidgets.QWidget):
             self.brush_size_input.clear()
         except ValueError:
             pass
-        self.brush.color = self.sidebar.brush.color
-    
+
+    def choose_color(self):
+        color = QtWidgets.QColorDialog.getColor()
+        if color.isValid():
+            self.brush.color = color
+            self.brush.set_color(color)
+            self.canvas.brush.color = color
